Extended Renderer::Context with an contextType property (WIN32, X11, WAYLAND). Now we support platform/windowing specific contexts. Added an X11 context implementation which holds a pointer to the X11 display connection
Extended Renderer::Context with an property for the renderer api shared library handle (e.g. optained by dlopen)
This property is needed/used under X11, in conjunction with an X11Context, because the OpenGL library (libGL.so) interacts with the X11 library (libX11.so), it registers some callbacks in the X11 library on load.

So the OpenGL library unload must be handled from the frontend side because only there we can satisfy the reguirements that the OpenGL Library gets unloaded after an call to XCloseDisplay
See http://dri.sourceforge.net/doc/DRIuserguide.html "11.5 libGL.so and dlopen()" for more details.

Extended RendererInstance class with support to load the renderer api library and set its handle value to the context
Added a method to the RendererInstance class to destroy the renderer instance explicitly, instead of only in the destructor.

The OpenGLRenderer::OpenGLContextLinux takes now an Renderer::Context instance instead of an window handle and an bool if external context is used or not
OpenGLRenderer::OpenGLRuntimeLinking now uses the handle value (if set) for the renderer api shared library instead of loading the library itself